commit_dec_13_blueprint_reg
