+++ app/reg/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+from flask_simple_captcha import CAPTCHA
+
+bp = Blueprint('reg', __name__)
+
+
+from app.reg import forms, routes
