+++ app/templates/reg/edit_profile.html
