+++ app/templates/reg/email/end_test.txt
