+++ app/templates/reg/email/forget_pwd_email.txt
