+++ app/templates/reg/email/register_email.html
