+++ app/templates/reg/email/register_email.txt
