+++ app/templates/reg/login.html
@@ -1,0 +1,35 @@
+{% extends"template.html" %}
+
+{% block scripts %}
+{% endblock %}
+
+{% block title %}
+    {{title}}
+{% endblock %}
+
+{% block content %}
+    <main class="container-fluid">
+        <div class="row">
+            <div class="col-3"></div>
+            <div class="col-3">
+                <h1>Авторизация на сайте</h1>
+                <form action="" method="post" novalidate>
+                    {{ form.hidden_tag() }}
+                    <p>
+                        {{ form.username.label }}<br>
+                        {{ form.username(size=32) }}
+                    </p>
+                    <p>
+                        {{ form.password.label }}<br>
+                        {{ form.password(size=32) }}
+                    </p>
+                    <p>{{ form.submit() }}</p>
+                </form>
+                <a href="{{ url_for('forget_pwd') }}">Забыли пароль?</a>
+            </div>
+            <div class="col-3"></div>
+            <div class="col-3"></div>
+            </div>
+        </div>
+	</main>
+{% endblock %}
