+++ app/templates/reg/register.html
