+++ app/templates/reg/send_pwd.html
